refactor(zkvm-test): move sp1 tests to zkvm test
