+++ provers/tests/Cargo.toml
@@ -9,28 +9,22 @@ bincode.workspace = true
 bitcoin.workspace = true
 borsh.workspace = true
 hex.workspace = true
+num-bigint = "0.4.6"
+num-traits = "0.2.19"
 serde_json.workspace = true
+sp1-sdk = "3.0.0"
 strata-primitives.workspace = true
 strata-proofimpl-btc-blockspace.workspace = true
 strata-proofimpl-checkpoint.workspace = true
-strata-proofimpl-cl-agg.workspace = true
 strata-proofimpl-cl-stf.workspace = true
 strata-proofimpl-evm-ee-stf.workspace = true
 strata-proofimpl-l1-batch.workspace = true
+strata-proofimpl-cl-agg.workspace = true
 strata-state.workspace = true
 strata-test-utils.workspace = true
 strata-zkvm.workspace = true
 tempfile.workspace = true
-
-# sp1
-strata-sp1-adapter = { workspace = true, optional = true }
-strata-sp1-guest-builder = { path = "../sp1", optional = true }
-
-# risc0
-strata-risc0-adapter = { workspace = true, optional = true }
-strata-risc0-guest-builder = { path = "../risc0", optional = true }
-
-[features]
-sp1 = ["strata-sp1-adapter", "strata-sp1-guest-builder"]
-risc0 = ["strata-risc0-adapter", "strata-risc0-guest-builder"]
-default = ["sp1"]
+# strata-sp1-adapter = { workspace = true, optional = true }
+# strata-sp1-guest-builder = { workspace = true, optional = true }
+strata-sp1-adapter = { workspace = true }
+strata-sp1-guest-builder = { path = "../sp1" }
