+++ provers/tests/src/l1_batch.rs
@@ -11,7 +11,7 @@ use strata_zkvm::{
     AggregationInput, Proof, ProofType, VerificationKey, ZkVmHost, ZkVmInputBuilder, ZkVmProver,
 };
 
-use crate::helpers::{btc::BtcBlockProofGenerator, proof_generator::ProofGenerator};
+use crate::{btc::BtcBlockProofGenerator, proof_generator::ProofGenerator};
 
 pub struct L1BatchProofGenerator {
     btc_proof_generator: BtcBlockProofGenerator,
@@ -81,6 +81,28 @@ mod test {
 
     #[test]
     fn test_l1_batch_code_trace_generation() {
+        let params = gen_params();
+        let rollup_params = params.rollup();
+        let l1_start_height = (rollup_params.genesis_l1_height + 1) as u32;
+        let l1_end_height = l1_start_height + 2;
+
+        let btc_proof_generator = BtcBlockProofGenerator::new();
+        let _ = L1BatchProofGenerator::new(btc_proof_generator)
+            .get_proof(&(l1_start_height, l1_end_height))
+            .unwrap();
+    }
+}
+
+// #[cfg(all(feature = "prover", not(debug_assertions)))]
+mod test {
+    use strata_test_utils::l2::gen_params;
+
+    use crate::{BtcBlockProofGenerator, L1BatchProofGenerator, ProofGenerator};
+
+    #[test]
+    fn test_l1_batch_code_trace_generation() {
+        sp1_sdk::utils::setup_logger();
+
         let params = gen_params();
         let rollup_params = params.rollup();
         let l1_start_height = (rollup_params.genesis_l1_height + 1) as u32;
